Dynamically generate backgrounds from giphy.
For now, I'm just pulling in a random background using giphy's
random endpoint, which was very easy to use. Configuration of the
API key is really janky for now, and will maybe be mopped up in
the future. I just hope that Giphy and X11R5 don't start banning
me.