Add redis and start saving joke data.
'channel' is just hardcoded for now, as there won't be a reason to
use more than one channel for a while. This is a pretty weird way
to use redis I bet (I'm basically making it behave like a
relational database), but, well, I'm an idiot and I'm sure as heck
not going to install python mysql things. I hate those.