Fix longstanding off-by-one error.
This had been ripping an innocent word out of the nonsense that
X11R5 generates. What a disaster! X11R5 spits out things that
do not really make sense, but when you go dropping a word out
right in the middle of the sentence, they only get harder to
comprehend.

The nonsense is now 80% more sensible, and hopefully 100% funnier.