Refactoring to kill globals.
Things have to be moved around here so that 'save_joke' exists at
the point that it is being called. Everything works here, so this
is getting committed before we move on.

This is part of a series of commits aimed at making views
responsible for a minimal amount of logic, and making jokes easy
to retrieve by identifier.