Kill the last global variable.
Jokes are now retained as long as the server is up.

Resets now just straight-up generate a new joke, instead of making
the "last_time" be further back in time than it really was -- which
has the wonderful benefit of putting the reset load on me as the
admin, instead of on the next viewer to hit the page after I
trigger a reset.

The time logic is also pretty isolated now, and might get broken
into a utility function.