Automatically refresh the main view page.
Individual ID pages shouldn't refresh and ride along the wave of
jokes, but the straight up 'view' page should do that. It
refreshes once a minute for now -- a better approach would be to
advance forward every fifteen seconds or so as long as another
joke is there to advance to, but, I'm pretty lazy at this point.